Fix cmake segfault by renaming ext_name_cxxopts
Our builds are failing with a segfault with cmake 3.26.5 on RHEL. The
final phase of CMake is trapped in an infinitive recursion and
eventually exhausts all stack trace. The segfault only occurs with
`USE_SYSTEM_DEPS=ON`.

Eventually we found the culprit in `extlib_cxxopts.cmake`. The line
`add_dependencies(${ext_name_cxxopts} cxxopts)` is causing the segfault.
`ext_name_cxxopts` is `cxxopts`. I'm not entirely sure why it is
failing. I think there is a variable name clash. Setting
`ext_name_cxxopts` to a different value resolves the problem.

Signed-off-by: Christian Heimes <christian@python.org>