Update pdf_parser.py
My project needs a font size (if available) within the pdf spans. It appears the C++ already contains this information, so we added it to the class.

The change ensures that if the `font-size` key exists in the data provided by the C++ layer, its value is read and passed during the creation of the `PdfTextCell` object. 

Note we also changed docling-core/docling_core/types/doc/page.py
to include an optional `font_size` field. This is reflected in a separate pull request.

Signed-off-by: Christopher Hoover <cahoover@gmail.com>